# bytes: speed up Replace